Make this somewhat less crazy, but actually still batshit insane really :(
The way that this previously worked was to parse the string argument as a type constraint, which resulted in you getting
class_type(20), i.e. registering a type constraint called '20' which then checked blessed($_) eq '20'.
However, that code was never ever called. Instead, the type constraint was consumed whole by stringification, resulting
in the type check generated by the constraint generator to pass, entirely by accident.

The unfortunate overloading of type constraint objects was removed in Moose 1.06, which of course broke this
as it was now checking length < 20=HASH(0x800d80) which doesn't work so good :)

I have changed the implementation to generate a type called '20' in the global type constraint registry
(ewwww!), which checks the length.. This type is then used to parameterize Varchar, and everything once
again works (although note that the error message generated has changed to became more standard)..

Also added checking and an exception for the value of the type parameter, I have no idea what this did previously,
but it at least throws some sort of sane error now..

I've left the docs intact (and ergo now wrong as the error has changed), I'll let you fix that up ;)
I hadn't really thought about this and its crapping all over the global type constraint registry before,
and that makes me sad panda.. :( I'll see if I can fiddle a bit more and come up with a way to get round
that without the errors saying Varchar[__ANON__]