ensure that warnings are not thrown fixes issue #2
in some programs a warning will cause an exception, and the warning from the
moose type will display first over the message itself. This can be quite
confusing when determing the cause of problems. This code will handle
undefined values so that they do not throw warnings when printing the message.

Signed-off-by: Caleb Cushing <xenoterracide@gmail.com>